Improved SubFixer Capabilities
Now SubFixer Can accept batch files and folders, and check if they are
subtitle, fix their encoding, and if they are folder fix their inside
subtitle file if there is.
Also, now app checks if the subtitle file, is in true encode ignores it
and goes to next file checking if there.